requirements/test.txt
pytest
pytest-django
pytest-mock
psycopg[binary,pool]
